kadai_tasklist

+++ app/Http/Controllers/TasksController.php
@@ -14,9 +14,16 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->get();
+            
+            return view('tasks.index',['tasks' => $tasks,]);    
+            
+        }
         
-        return view('tasks.index',['tasks' => $tasks,]);
+        return view('welcome');
+        
     }
 
     /**
@@ -44,10 +51,10 @@ class TasksController extends Controller
             'content' => 'required | max:191',
             ]);
             
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content, 
+            'status' => $request->status,
+            ]);
         
         return redirect('/');
     }
@@ -61,8 +68,11 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
-        
-        return view('tasks.show',['task' => $task,]);
+        if (\Auth::ID === $task->user_id) {
+            $task->delete();
+        }
+        //return view('tasks.show',['task' => $task,]);
+        return redirect('/');
     }
 
     /**
@@ -74,8 +84,12 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::find($id);
+        if (\Auth::ID === $task->user_id) {
+            $task->delete();
+        }
         
-        return view('tasks.edit',['task' => $task,]);
+        //return view('tasks.edit',['task' => $task,]);
+        return redirect('/');
     }
 
     /**
@@ -93,6 +107,10 @@ class TasksController extends Controller
             ]);
             
         $task = Task::find($id);
+        if (\Auth::ID === $task->user_id) {
+            $task->delete();
+        }
+        
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -109,7 +127,9 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        if (\Auth::ID === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
